refactor:  `interface` 추가
